refactor: update import logic across the package
