+++ pydriosm/downloader/__init__.py
@@ -3,9 +3,8 @@ Download `OpenStreetMap <https://www.openstreetmap.org/>`_ (OSM) data from free 
 `Geofabrik <https://download.geofabrik.de/>`_ and `BBBike <https://download.bbbike.org/>`_.
 """
 
-from . import _base, web_parser
 from ._bbbike import BBBikeDownloader
 from ._geofabrik import GeofabrikDownloader
 from ._wrapper import Downloader
 
-__all__ = ['BBBikeDownloader', 'GeofabrikDownloader', 'Downloader', '_base', 'web_parser']
+__all__ = ['BBBikeDownloader', 'GeofabrikDownloader', 'Downloader']
