+++ pydriosm/ios/__init__.py
@@ -2,9 +2,8 @@
 Implement storage I/O of (parsed) OSM data extracts with `PostgreSQL <https://www.postgresql.org/>`_.
 """
 
-from . import _base, utils
 from ._bbbike import BBBikeIOS
 from ._geofabrik import GeofabrikIOS
 from ._pgsql_osm import PostgresOSM
 
-__all__ = ['PostgresOSM', 'GeofabrikIOS', 'BBBikeIOS', '_base', 'utils']
+__all__ = ['PostgresOSM', 'GeofabrikIOS', 'BBBikeIOS']
